requirements.txt
pyyaml>=6.0
python-terraform>=0.10.1
google-cloud-compute>=2.0.0